Refactor to use the FormatterFactory interface
This a major API change that will allow to build much more performant
formatters.
The change will affect the way to write Formatters, as formatters now
have a constructor (a factory) that is able to return the best
Formatter implementation based on user settings.
The immediate benefit implemented in this commit is that the
TextFormatter will now check if the output writer is a terminal just
once at initialisation, while this check was previously done ON EVERY
LOG LINE.
JSONFormatter and LogstashFormatter are also improved as their default
timestamp format is set just once at init instead of on every log line.

This API change implemented in a way so that few external code will be
affected. For example, the following lines will not have to be changed
despites the SetFormatter method signature changes:
    log = logrus.New()
    log.SetFormatter(&TextFormatter{})